Update Writing-Addon doc
(two line change)
Update Writing-Addon doc to fix error in reading parameter data.  As currently written this causes an exception due to the object returned not being a string. Two ways to fix this, change the data passed by param to just a bare string in the story code above, or change how the addon handles the parameter data. The second is preferable since this is the more likely scenario, and a more useful showcase of passing parameter data.   This is the path I've chosen to changed.