Changes import of options/optionsKnob
I run into the error mentioned here: https://github.com/storybookjs/storybook/issues/7116

The import, which is different from all other imports, is easy to overlook. Therefore I have adapted the import to the other button types and added a note for an alternative import that corresponds to the naming of the other button types (without the suffix `Knob`).