fix: update the command used to get package's version using CLI with Yarn 2
Using CLI with Yarn 2 and Node 10 results in an output polluted with:
`(node:87281) ExperimentalWarning: The XXX.promises API is experimental`

These messages are warnings but are sent to `stderr` which is causing CLI to throw an error too.
With `sync` function we can check the `status` to know if command ends in error or not and process accordingly.