test(cli): don't test CLI against `react_scripts` fixtures with Yarn 2
When testing Yarn 2 compatibility of Storybook CLI, we are using Yarn 2 in fixture directories inside a Yarn 1 workspace.
To access local versions of `@storybook` packages we use Yarn 2 `portal` protocol (see berryfy_package_json.js). However, it looks like CRA is not playing well with `portal` protocol and is throwing this kind of error: `Relative imports outside of src/ are not supported.`.

As new E2E tests will be soon available, I just removed the CRA tests instead of trying to hack them a bit more.