refactor: replace `semver` by `@storybook/semver`
`@storybook/semver` is a browser friendly fork of `semver`.
We use it to avoid issue with IE11 as classic "compiled" `semver` is now compatible with it.
If we just recompile `semver` from `node_modules` it ends with a package that `require('core-js')` but `core-js` is not listed in the dependency. And so Yarn 2 throw an error.
So to avoid that we need to use a "proper" dependency which is `semver` recompile and with all needed dependencies in its package.json.