fix(essentials): use `require.resolve` to ensure Yarn PnP compatibility
Files of various addons should be resolved in the context of `addon-essentials` and not in the one of SB user projects.
File to load can be `preset.js`, `register.js`, or the package entry point, so we need to check all these cases as it's done in `lib/core/src/server/presets.js`.