Fix copy for Docs
The website state that DocsPage is available out of the box with Storybook. After upgrading to Storybook v6 in a recent project, I didn't have access to Docs, and had to install the addon separately. You can also see in the [next page](https://storybook.js.org/docs/react/writing-docs/docs-page) that it states you need to install Storybook Docs to get access to DocsPage (which is more accurate, albeit conflicting and confusing in it's present state).

I changed the copy from `Out of the box` to `When you install Storybook Docs` to clarify that the user does indeed need to install the addon to access the functionality.