chore: set a "fixed" version of Aurelia
`dev` tag was used for `aurelia` dep in both:
 - `@storybook/aurelia`
 - `aurelia-kitchen-sink`

A new version is published with this tag multiples a week.
But there is no real benefit for us to use a non "stable" version.
So I set the version to the 0.7.0 (latest stable as of today).

If we need/want to test against `dev` version of aurelia when can setup a specific E2E test.