Move `react` and `react-dom` to peer deps
`react` and `react-dom` should always be peer dependencies of packages that provide react components to ensure that there's a single instance of those.

Plus, all packages depending on a package with `react` as peerDep should also have it as peerDeps (transitive peer dependencies) or as deps.

Finally, packages in `examples/*` are "end" packages using Storybook so they must provide `react` and `react-dom` in their deps/devDeps