fix(addon-docs): Do not create additional Vue tree
This commit changes how the decorator grab Vue.js' VNode tree.
Previously it creates a new Vue instance and traverse the tree.
But this causes a lot of problems such as unexpected mount call and
DOM-detached instance. I removed the instruction then added a decorator
that do the instruction from inside of an existing VNode tree.